requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
graphdatascience>=1.0.0
neo4j>=5.0.0
google-generativeai>=0.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
langchain>=0.1.0
langchain-google-genai>=0.0.8
langgraph>=0.0.10
google-generativeai>=0.3.0
gradio>=4.0.0
pydantic>=2.0.0
